requirements.txt
Brotli
certifi
click
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-table
Flask
Flask-Compress
influxdb-client
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
plotly
pyserial
python-dateutil
pytz
Rx
six
tenacity
urllib3
Werkzeug
flightsql-dbapi